Better command line interface. Holy fuck that was hard as fuck.
But now it's super sexy. Readme updated with new command. We use subcommands so the user can specify what they would like to do. The main one is optimize. There is also generate and hash which will generate a supercuts.json template file (based on the input root files); dump the cuts applied for the given hash(es) specified.